requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
anyio==4.9.0
av==14.2.0
black==25.1.0
blinker==1.9.0
blis==1.2.0
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
decord==0.6.0
dill==0.3.8
distro==1.9.0
einops==0.8.0
executing==2.2.0
fairscale==0.4.13
ffmpy==0.5.0
ffprobe==0.5
ffprobe-python==1.0.3
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
func_timeout==4.3.5
future==1.0.0
google-auth==2.39.0
google-genai==1.11.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
icecream==2.1.4
idna==3.10
importlib_resources==6.5.2
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lazy_loader==0.4
lightning-utilities==0.14.3
llava @ git+https://github.com/LLaVA-VL/LLaVA-NeXT.git@09e5840d5589ad2d6a8656c0a60f21ae134b3309
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mutagen==1.47.0
mypy-extensions==1.0.0
narwhals==1.31.0
ninja==1.11.1.1
nltk==3.9.1
omegaconf==2.3.0
openai==1.14.2
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathos==0.3.2
pathspec==0.12.1
pillow==11.1.0
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
prettytable==3.10.0
propcache==0.2.1
protobuf==3.20.0
psutil==6.1.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.8
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
ratelimit==2.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
ruff==0.11.0
safetensors==0.5.3
SceneGraphParser==0.1.0
scikit-image==0.25.2
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
squarify==0.4.4
srsly==2.5.1
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.3.30
tiktoken==0.6.0
timm==1.0.14
tokenizers==0.19
tomlkit==0.12.0
tqdm==4.67.1
transformers==4.41.0
triton==3.1.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
